Moving clientauth package to be under client
